\name{MAMSNews}
\alias{MAMSNews}
\title{Shows changes and news}
\description{Functions showing changes since previous versions.}
\usage{
MAMSNews()
}

\details{
Displays the changes and news given in the NEWS file of the package.
}

\value{Screen output.}

\author{Thomas Jaki}

\examples{
MAMSNews()
}

\keyword{misc}
